grimoirelab-api/requirements.txt
flask==2.0.1
werkzeug==2.0.1
requests==2.26.0
python-dotenv==0.19.0
docker==5.0.0
six>=1.10.0
gitpython==3.1.30
elasticsearch>=7.0.0,<8.0.0
PyGithub>=1.55
networkx>=2.6.3
numpy>=1.21.0
scipy>=1.7.0